@@ -15,13 +15,14 @@
   },
   "dependencies": {
     "fastclick": "1.0.6",
-    "react": "15.4.2",
-    "react-dom": "15.4.2",
+    "prop-types": "^15.5.10",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
     "react-router": "3.0.2",
-    "styled-components": "1.3.1"
+    "styled-components": "1.4.6"
   },
   "devDependencies": {
-    "gh-pages": "^0.12.0",
-    "react-scripts": "0.8.5"
+    "gh-pages": "1.0.0",
+    "react-scripts": "1.0.2"
   }
 }
