@@ -1,4 +1,5 @@
 import React from "react";
+import PropTypes from "prop-types";
 import withResponsiveness
   from "../../higher-order-components/withResponsiveness";
 import HomeMobile from "./HomeMobile.js";
@@ -7,7 +8,7 @@ import HomeDesktop from "./HomeDesktop.js";
 const Home = ({ isMobile }) => isMobile ? <HomeMobile /> : <HomeDesktop />;
 
 Home.propTypes = {
-  isMobile: React.PropTypes.bool.isRequired,
+  isMobile: PropTypes.bool.isRequired,
 };
 
 export default withResponsiveness(Home);
